🎨 pages, components 디렉토리 구조 변경
- /components/commons -> /components로 대체
- /pages 하위로 기존 /components의 페이지 컴포넌트 폴더 이동
