+++ app/frontend/src/components/Chatting/ChattingFooter.tsx
