+++ app/frontend/src/components/Chatting/TalkItem.css.ts
