+++ app/frontend/src/components/Chatting/TalkItem.tsx
