+++ app/frontend/src/components/Error/index.css.ts
