+++ app/frontend/src/components/Error/index.tsx
