+++ app/frontend/src/components/Label/index.css.ts
