+++ app/frontend/src/components/Label/index.tsx
