+++ app/frontend/src/components/Modal/Footer.tsx
