+++ app/frontend/src/components/Modal/index.css.ts
