+++ app/frontend/src/components/MogacoItem/index.css.ts
