+++ app/frontend/src/components/MogacoItem/index.tsx
