+++ app/frontend/src/components/Sidebar/index.css.ts
