+++ app/frontend/src/components/Sidebar/index.tsx
