+++ app/frontend/src/components/UserChip/index.tsx
