+++ app/frontend/src/components/index.ts
@@ -1,17 +1,14 @@
-export * from './Mogaco';
-export * from './MogacoDetail';
-export * from './MogacoPost';
-export * from './commons/Button';
-export * from './commons/Chatting';
-export * from './commons/Error';
-export * from './commons/Header';
-export * from './commons/Input';
-export * from './commons/Input/Textarea';
-export * from './commons/Label';
-export * from './commons/Layout';
-export * from './commons/Loading';
-export * from './commons/Loading/LoadingIndicator';
-export * from './commons/MogacoItem';
-export * from './commons/Popover';
-export * from './commons/Sidebar';
-export * from './commons/UserChip';
+export * from './Button';
+export * from './Chatting';
+export * from './Error';
+export * from './Header';
+export * from './Input';
+export * from './Input/Textarea';
+export * from './Label';
+export * from './Layout';
+export * from './Loading';
+export * from './Loading/LoadingIndicator';
+export * from './MogacoItem';
+export * from './Popover';
+export * from './Sidebar';
+export * from './UserChip';
