+++ app/frontend/src/pages/Mogaco/MogacoList.css.ts
