+++ app/frontend/src/pages/Mogaco/MogacoList.tsx
