+++ app/frontend/src/pages/Mogaco/index.tsx
@@ -1,5 +1,12 @@
-import { MogacoPage } from '@/components';
+import * as styles from './index.css';
+import { MogacoList } from './MogacoList';
+import { MogacoListHeader } from './MogacoListHeader';
 
-export function Mogaco() {
-  return <MogacoPage />;
+export function MogacoPage() {
+  return (
+    <div className={styles.container}>
+      <MogacoListHeader />
+      <MogacoList />
+    </div>
+  );
 }
