+++ app/frontend/src/pages/MogacoDetail/DetailHeader.tsx
