+++ app/frontend/src/pages/MogacoDetail/DetailHeaderButtons.tsx
