+++ app/frontend/src/pages/MogacoDetail/DetailInfo.tsx
