+++ app/frontend/src/pages/MogacoDetail/index.css.ts
