+++ app/frontend/src/pages/MogacoPost/MogacoPostTitle.tsx
